Add Streamlit chatbot, Streamlit RAG chat, llama2 RAG  and Langchain fundamentals
pages/Chatbot.py -> Add a chatbot that utilizes memory retrieval to maintain context and continuity in conversations

pages/File Chat.py -> Add a chatbot that employs RAG to generate responses based on the content of uploaded PDF files.

notebooks/rag_llama2.ipynb -> Add a notebook that focuses on the application of RAG using a local GPU with the llama2 model.

notebooks/langchain_concepts.ipynb -> Add a file with fundamental concepts of the LangChain library,